fix path

+++ bin/npmt.js
@@ -15,16 +15,16 @@ install.on('close', code => {
 	stats.averageTime = newAvg;
 	stats.installCount++;
 	fileUtils.writeStats(stats);
-  if (countdown) {
-    process.kill(-countdown.pid);
-  }
+	if (countdown && countdown.connected) {
+		process.kill(-countdown.pid);
+	}
 });
 
 if(fileUtils.isHasStats()) {
 	let seconds = Math.floor(stats.averageTime / 1000);
 	const min = utils.zeroPad(Math.floor(seconds / 60), 2);
 	seconds = utils.zeroPad(seconds - min * 60, 2);
-	countdown = spawn('./bin/countdown.sh', [`00:${min}:${seconds}`], {stdio: 'inherit', detached: true});
+	countdown = spawn(`${__dirname}/countdown.sh`, [`00:${min}:${seconds}`], {stdio: 'inherit', detached: true});
 
 } else {
 	console.log('No npm stats yet. wait for next time ;)')
